+++ CHANGELOG.md
@@ -25,6 +25,7 @@ dev
    - Improve GGA performance
    - Do an unpaired calculation automatically if the system is unpaired
    - Option to use a symmetric initial guess, i.e., the same guess for both spin channels
+   - Add trajectory keyword to XYZ and PDB writer to append geometries
    - New functions for the
       - Electron localization function (ELF)
       - Positive-definite kinetic energy density